Workarounds for last commit:
Discovered painfully that shortcutted and statements seem to be broken or not actually a thing in CL or something, so that whole 'gate of setqs' thing i had earlier doesn't work (which i was taught would work in University).
This happened when i did a drop-in replacement of these functions to an existing project that used them extensively from alexandria.
The debugger stopped at a crash midway through x-let* varients and reported that the earlier bindings were nil and should have been stopped by the AND statement.
This probably means the and statement in SBCL evaluates its arguments in parallel or something.
So now instead i have a tower of setqs, and that hasn't seemingly had any problem yet.